fix(ci,controller): drop deprecated Node.js 20 actions; remove useless i32::from
Two CI failures on PR #225 (issue-207):

## clippy: useless `i32::from(i32)`

`Gateway.spec.listeners[].port` from the gateway-api crate is already `i32`;
the `i32::from(listener.port)` calls in `render::service_ports` and
`render::container_ports` were no-ops that `clippy::useless_conversion`
rightly rejects under `-D warnings`. Dropped.

## GitHub Actions: bump Node-20 pins to Node-24 majors

Audited every action referenced under `.github/workflows/`. Bumps below are
to the floating major tag (upstream publishes one in every case here);
mixed-version usage across workflow files consolidated to a single pin.

| Action | From | To | Reason |
|---|---|---|---|
| `dorny/paths-filter` | `@v3` | `@v4` | v3 runs node20 (deprecated); v4 is node24 |
| `azure/setup-helm` | `@v4` (distribution) + `@v5` (release) | `@v5` | v4 is node20; consolidate to v5 (node24) |
| `docker/build-push-action` | `@v6` (distribution) + `@v7` (release) | `@v7` | v6 is node20; consolidate to v7 (node24) |
| `docker/setup-buildx-action` | `@v3` (distribution) + `@v4` (release ×2) | `@v4` | v3 is node20; consolidate to v4 (node24) |
| `actions/download-artifact` | `@v7` | `@v8` | v7 is node24 (not deprecated) but v8 is the current supported major |

`.github/workflows/distribution.yml` was the laggard for three of these;
the rest were `actions/download-artifact` spread across `e2e.yml` and
`release.yml`.

### Upstream blocker (no action possible today)

- `actions/delete-package-versions@v5` still uses node20 at HEAD. No node24
  release upstream; the deprecation warning for this one will persist
  until the maintainers ship a new major. Tracked: re-audit on the next
  GitHub Actions deprecation cycle.

The other 14 action references in CI (`actions/cache@v5`,
`actions/checkout@v6`, `actions/labeler@v6`, `actions/setup-go@v6`,
`actions/upload-artifact@v7`, `aquasecurity/trivy-action@v0.36.0`,
`astral-sh/setup-uv@v8.2.0`, `docker/login-action@v4`,
`docker/metadata-action@v6`, `github/codeql-action/upload-sarif@v4`,
`helm/kind-action@v1`, `imranismail/setup-kustomize@v3`,
`sigstore/cosign-installer@v4.1.2`, `softprops/action-gh-release@v3`,
`taiki-e/install-action@v2`) are already on node24 or are
composite/docker actions with no Node runtime — no changes needed.

Refs #207

+++ crates/coxswain-controller/src/operator/render.rs
@@ -203,7 +203,7 @@ fn service_ports(gateway: &Gateway) -> Vec<ServicePort> {
     let mut seen: BTreeSet<(i32, &'static str)> = BTreeSet::new();
     let mut out = Vec::new();
     for listener in &gateway.spec.listeners {
-        let port = i32::from(listener.port);
+        let port = listener.port;
         let protocol = "TCP";
         if !seen.insert((port, protocol)) {
             continue;
@@ -304,7 +304,7 @@ fn container_ports(gateway: &Gateway) -> Vec<ContainerPort> {
     let mut seen: BTreeSet<i32> = BTreeSet::new();
     let mut out = Vec::new();
     for listener in &gateway.spec.listeners {
-        let port = i32::from(listener.port);
+        let port = listener.port;
         if !seen.insert(port) {
             continue;
         }
